third_party/blink/web_tests/platform/generic/fast/loader/simultaneous-reloads-assert-expected.txt

ALERT: If you do not hit an assertion failure when running this test with a debug build and you get a SUCCESS message, then you pass the test.
ALERT: SUCCESS
